# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Conditions API expose CRUD actions to condition objects on servers",
        "title": "Condition orchestrator API",
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
        "/serverEnroll/{uuid}": {
            "post": {
                "description": "Creates a server record in FleetDB and schedules an inventory condition on the device.\nIt will create a new server ID if UUID is not provided.\nSample server enroll request and response: https://github.com/metal-automata/conditionorc/blob/main/sample/serverenroll.md",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Server Enroll",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_metal-automata_conditionorc_pkg_api_v1_conditions_types.ServerResponse"
                        }
                    }
                }
            }
        },
...
        "/servers/{uuid}": {
            "delete": {
                "description": "Delete a server from FleetDB\nSample server delete request and response: https://github.com/metal-automata/conditionorc/blob/main/sample/serverenroll.md",
                "summary": "Server Delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_metal-automata_conditionorc_pkg_api_v1_conditions_types.ServerResponse"
                        }
                    }
                }
            }
        },
        "/servers/{uuid}/biosControl": {
            "post": {
                "description": "Controls the BIOS of the server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Bios Control",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "bios control options",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/condition.BiosControlTaskParameters"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_metal-automata_conditionorc_pkg_api_v1_conditions_types.ServerResponse"
                        }
                    }
                }
            }
        },
...
        "/servers/{uuid}/condition/{conditionKind}": {
            "post": {
                "description": "Creates a condition on a server\nSample firmwareInstall payload, response: https://github.com/metal-automata/conditionorc/blob/main/sample/firmwareInstall.md",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Condition Create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Condition Kind",
                        "name": "conditionKind",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_metal-automata_conditionorc_pkg_api_v1_conditions_types.ServerResponse"
                        }
                    }
                }
            }
        },
        "/servers/{uuid}/firmwareInstall": {
            "post": {
                "description": "Installs firmware on a device and validates with a subsequent inventory\nSample firmwareInstall payload, response: https://github.com/metal-automata/conditionorc/blob/main/sample/firmwareInstall.md",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Firmware Install",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "firmware install options",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/condition.FirmwareInstallTaskParameters"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_metal-automata_conditionorc_pkg_api_v1_conditions_types.ServerResponse"
                        }
                    }
                }
            }
        },
        "/servers/{uuid}/status": {
            "get": {
                "description": "Returns condition of a server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Condition Status",
                "operationId": "conditionStatus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_metal-automata_conditionorc_pkg_api_v1_conditions_types.ServerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "condition.BiosControlTaskParameters": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "The bios control action to be performed\n\nRequired: true",
                    "allOf": [
                        {
                            "$ref": "#/definitions/condition.BiosControlAction"
                        }
                    ]
                },
                "asset_id": {
                    "description": "Identifier for the Asset in the Asset store.\n\nRequired: true",
                    "type": "string"
                },
                "bios_config_url": {
                    "description": "The URL for the bios configuration settings file.\nNeeded for BiosControlAction.SetConfig\n\nRequired: false",
                    "allOf": [
                        {
                            "$ref": "#/definitions/url.URL"
                        }
                    ]
                }
            }
        },
...
        "condition.FirmwareInstallTaskParameters": {
            "type": "object",
            "properties": {
                "asset_id": {
                    "description": "Inventory identifier for the asset to install firmware on.",
                    "type": "string"
                },
                "dry_run": {
                    "description": "When defined, flasher will not perform any disruptive actions on the asset,\nit will download the firmware to be installed and determine if the firmware is applicable for the device.\n\nNo firmware installs will be attempted and if the device is powered off, it will not be powered on.",
                    "type": "boolean"
                },
                "firmware_set_id": {
                    "description": "FirmwareSetID specifies the firmware set to be applied.",
                    "type": "string"
                },
                "firmwares": {
                    "description": "Firmwares is the list of firmwares to be installed.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/condition.Firmware"
                    }
                },
                "force_install": {
                    "description": "Force install given firmware regardless of current firmware version.",
                    "type": "boolean"
                },
                "require_host_powered_off": {
                    "description": "When true, flasher will expect the host to be powered off before proceeding,\nif the host is not already powered off - the install task will be failed.",
                    "type": "boolean"
                },
                "reset_bmc_before_install": {
                    "description": "Reset device BMC before firmware install",
                    "type": "boolean"
                }
            }
        },
...
        "github_com_metal-automata_conditionorc_pkg_api_v1_conditions_types.ServerResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "records": {
                    "$ref": "#/definitions/types.ConditionsResponse"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "types.ConditionsResponse": {
            "type": "object",
            "properties": {
                "conditions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/condition.Condition"
                    }
                },
                "serverID": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/condition.State"
                }
            }
        },
...
    }
}